added the add to wishlist functionality and dummy remove from wishlist functionality

+++ app/context/CartContext.js
@@ -1,5 +1,6 @@
 "use client";
 
+import { onAuthStateChanged } from "firebase/auth";
 import { auth } from "../lib/firebaseconfig";
 
 const { createContext, useState, useEffect } = require("react");
@@ -9,13 +10,37 @@ export const CartContext = createContext();
 export function CartProvider({ children }) {
   const [cartItems, setCartItems] = useState([]);
   const [wishlist, setWishlist] = useState([]);
-  const [wishlisted, setWishlisted] = useState(false);
   useEffect(() => {
     const storedData = localStorage.getItem("cart");
     if (storedData) {
       setCartItems(JSON.parse(storedData));
     }
   }, []);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      async function fetchWishlist() {
+        if (!user) return;
+
+        try {
+          const res = await fetch(`/api/getwishlist?uid=${user.uid}`);
+          if (!res.ok) {
+            throw new Error("Failed to fetch wishlisted products");
+          }
+
+          const data = await res.json();
+          console.log("Fetched wishlists successfully ");
+          setWishlist(data);
+        } catch (error) {
+          console.error("Failed to fetch wishlist", error.message);
+        }
+      }
+
+      fetchWishlist();
+    });
+
+    return () => unsubscribe();
+  }, [auth.currentUser]);
 
   useEffect(() => {
     localStorage.setItem("cart", JSON.stringify(cartItems));
@@ -85,35 +110,63 @@ export function CartProvider({ children }) {
       }
       setWishlist((prev) => {
         const existingItem = prev.find((item) => item.id === selectedItem.id);
-        let updatedWishlist;
 
-        updatedWishlist = existingItem
-          ? prev.map((item) =>
-              item.id === selectedItem.id
-                ? { ...item, quantity: item.quantity + 1 }
-                : item
-            )
-          : [...prev, { ...selectedItem, quantity: 1 }];
+        if (existingItem) {
+          return prev;
+        }
+
+        const updatedWishlist = [...prev, { ...selectedItem, quantity: 1 }];
 
         localStorage.setItem("wishlist", JSON.stringify(updatedWishlist));
         return updatedWishlist;
       });
-
-      setWishlisted(true);
     } catch (error) {
       console.error("Error adding product to wishlist", error.message);
+    }
+  }
+
+  async function deleteWishlist(Item) {
+    const user = auth.currentUser;
+
+    if (!Item) {
+      console.log("Item invalid, please try again!");
+    }
+
+    try {
+      const res = await fetch("/api/removewishlist", {
+        method: "DELETE",
+        headers: { "Content-type": "application/json" },
+        body: JSON.stringify({
+          uid: user.uid,
+          id: Item.id,
+        }),
+      });
+
+      if (!res.ok) {
+        throw new Error("Error deleting item from wishlist");
+      }
+
+      setWishlist((prev) => {
+        const updatedData = prev.filter((item) => item.id !== Item.id);
+        localStorage.setItem("wishlist", JSON.stringify(updatedData));
+        return updatedData;
+      });
+      console.log("Item successfully removed from wishlist.");
+    } catch (error) {
+      console.error("Error removing product from wishlist", error.message);
     }
   }
 
   return (
     <CartContext.Provider
       value={{
-        wishlisted,
+        wishlist,
         cartItems,
         setCartItems,
         addToCart,
         removeFromCart,
         addToWishlist,
+        deleteWishlist,
       }}
     >
       {children}
